dashboard/requirements.txt
streamlit
PyMuPDF
pytesseract
Pillow
pandas
qdrant-client
langchain-openai
python-dotenv
langchain
